[cheri] Update objectHandle/object_handle params to uintptr_t

+++ layers/object_tracker/object_lifetime_validation.cpp
@@ -1798,7 +1798,7 @@ bool Device::PreCallValidateSetPrivateData(VkDevice device, VkObjectType objectT
 }
 
 #if defined(__CHERI_PURE_CAPABILITY__)
-bool Device::PreCallValidateGetPrivateData(VkDevice device, VkObjectType objectType, uint64_t objectHandle,
+bool Device::PreCallValidateGetPrivateData(VkDevice device, VkObjectType objectType, uintptr_t objectHandle,
 #else   // !__CHERI_PURE_CAPABILITY__
 bool Device::PreCallValidateGetPrivateData(VkDevice device, VkObjectType objectType, uint64_t objectHandle,
 #endif  // !__CHERI_PURE_CAPABILITY__
